add: implementado a requisição para estatisticas de load da CPU
