+++ src/routes/cpuLoadRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const cpuLoadController = require('../controllers/cpuLoadController');
+
+const router = express.Router();
+
+router.get('/', cpuLoadController.getCpuLoadInfo);
+
+module.exports = router;
